backend/requirements.txt
asgiref==3.2.10
autopep8==1.5.6
defusedxml==0.7.1
Django==3.0.5
django-colorfield==0.4.2
django-filter==2.3.0
djangorestframework==3.11.0
djoser==2.1.0
filters==1.3.2
flake8==3.9.0
gunicorn==20.1.0
Pillow==8.3.1
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2019.3
regex==2021.3.17
reportlab==3.6.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1